pyproject.toml
[tool.poetry]
name = "movie-repository"
version = "0.1.0"
description = ""
authors = ["Dioxide_CN <dioxide.cn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
requests = "^2.32.2"
bs4 = "^0.0.2"
lxml = "^5.2.2"
aiohttp = "^3.9.5"
motor = "^3.4.0"
pyyaml = "^6.0.1"
aiofiles = "^23.2.1"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
